+++ satella/coding/predicates.py
@@ -9,14 +9,23 @@ def _nop(v):
 
 
 def make_operation_two_args(operation_two_args: tp.Callable[[tp.Any, tp.Any], tp.Any],
-                            docstring: tp.Optional[str] = None):
+                            docstring: tp.Optional[str] = None,
+                            swap_order: bool = True):
     def operation(self, a) -> 'Predicate':
         if isinstance(a, Predicate):
-            def op(v):
-                return operation_two_args(self(v), a(v))
+            if not swap_order:
+                def op(v):
+                    return operation_two_args(self(v), a(v))
+            else:
+                def op(v):
+                    return operation_two_args(a(v), self(v))
         else:
-            def op(v):
-                return operation_two_args(self(v), a)
+            if not swap_order:
+                def op(v):
+                    return operation_two_args(self(v), a)
+            else:
+                def op(v):
+                    return operation_two_args(a, self(v))
         return Predicate(op)
     if docstring:
         operation.__doc__ = docstring
@@ -93,18 +102,31 @@ class Predicate:
     __add__ = make_operation_two_args(operator.add)
     __sub__ = make_operation_two_args(operator.sub)
     __mul__ = make_operation_two_args(operator.mul)
+    __matmul__ = make_operation_two_args(operator.matmul)
     __and__ = make_operation_two_args(operator.and_)
     __or__ = make_operation_two_args(operator.or_)
     __rshift__ = make_operation_two_args(operator.rshift)
     __lshift__ = make_operation_two_args(operator.lshift)
     __xor__ = make_operation_two_args(operator.xor)
+    __truediv__ = make_operation_two_args(operator.__truediv__)
+    __floordiv__ = make_operation_two_args(operator.floordiv)
+    __mod__ = make_operation_two_args(operator.mod)
+    __radd__ = make_operation_two_args(operator.add, swap_order=True)
+    __rsub__ = make_operation_two_args(operator.sub, swap_order=True)
+    __rmul__ = make_operation_two_args(operator.mul, swap_order=True)
+    __rmatmul__ = make_operation_two_args(operator.matmul, swap_order=True)
+    __rand__ = make_operation_two_args(operator.and_, swap_order=True)
+    __ror__ = make_operation_two_args(operator.or_, swap_order=True)
+    __rrshift__ = make_operation_two_args(operator.rshift, swap_order=True)
+    __rlshift__ = make_operation_two_args(operator.lshift, swap_order=True)
+    __rxor__ = make_operation_two_args(operator.xor, swap_order=True)
+    __rtruediv__ = make_operation_two_args(operator.__truediv__, swap_order=True)
+    __rfloordiv__ = make_operation_two_args(operator.floordiv, swap_order=True)
+    __rmod__ = make_operation_two_args(operator.mod, swap_order=True)
     __neg__ = make_operation_single_arg(operator.neg)
     __pow__ = make_operation_two_args(operator.pow)
     __invert__ = make_operation_single_arg(operator.invert)
     __abs__ = make_operation_single_arg(abs)
-    __truediv__ = make_operation_two_args(operator.__truediv__)
-    __floordiv__ = make_operation_two_args(operator.floordiv)
-    __mod__ = make_operation_two_args(operator.mod)
 
 
 x = Predicate()