+++ tests/test_coding/test_sequences.py
@@ -4,10 +4,16 @@ import unittest
 from satella.coding.sequences import choose, choose_one, infinite_counter, take_n, is_instance, \
     is_last, add_next, half_cartesian, skip_first, zip_shifted, stop_after, group_quantity, \
     iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, odd, Multirun, n_th, \
-    unique, length, map_list
+    unique, length, map_list, smart_zip
 
 
 class TestSequences(unittest.TestCase):
+
+    def test_smart_zip(self):
+        a1 = [(1, 1), (1, 2), (1, 3)]
+        a2 = [1, 2, 3]
+        a = list(smart_zip(a1, a2))
+        self.assertEqual(a, [(1, 1, 1), (1, 2, 2), (1, 3, 3)])
 
     def test_map_list(self):
         _list = map_list(lambda x: x*2, range(5))