+++ tests/test_coding/test_predicates.py
@@ -6,6 +6,10 @@ from satella.coding.predicates import x
 
 class TestPredicates(unittest.TestCase):
 
+    def test_upper(self):
+        p = x.upper()
+        self.assertEquals(p('ala ma kota'), 'ALA MA KOTA')
+
     def test_instanceof(self):
         p = x.instanceof(int)
         self.assertTrue(p(2))