+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.11.1
+
+* added extra operators to `Predicate`
+