+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 * added extra operators to `Predicate`
 * deprecated `skip_first` and `stop_after`
-
+* added extra docs for iterators