+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * added extra operators to `Predicate`
 * deprecated `skip_first` and `stop_after`
 * added extra docs for iterators
+* added `map_list`