+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * deprecated `skip_first` and `stop_after`
 * added extra docs for iterators
 * added `map_list`
+* fix typing issues for `Predicate`