+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.2
+
+* added `smart_zip`