+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.11.2
 
 * added `smart_zip`
+* added `step` to `smart_enumerate`