+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * added `smart_zip`
 * added `step` to `smart_enumerate`
+* reserve more memory in case of MemoryError
+