+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 * added `smart_zip`
 * added `step` to `smart_enumerate`
 * reserve more memory in case of MemoryError
-
+* added `call_in_separate_thread`