+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * added `step` to `smart_enumerate`
 * reserve more memory in case of MemoryError
 * added `call_in_separate_thread`
+* added `read_nowait`