+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # v2.11.3
 
+* upgraded predicate to support no-arg invocations
+
+