+++ CHANGELOG.md
@@ -1,6 +1,4 @@
-# v2.11.3
-
-* upgraded predicate to support no-arg invocations
+# v2.11.4